Reverted Java 10 changes

+++ src/inception-scheduler/src/main/java/digital/inception/scheduler/SchedulerService.java
@@ -20,9 +20,9 @@ package digital.inception.scheduler;
 
 import digital.inception.core.util.ServiceUtil;
 import digital.inception.core.util.StringUtil;
+import digital.inception.validation.InvalidArgumentException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -31,6 +31,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.PostConstruct;
 import javax.sql.DataSource;
 import java.sql.*;
 import java.util.*;
@@ -44,9 +45,8 @@ import java.util.Date;
  * @author Marcus Portmann
  */
 @Service
-@SuppressWarnings("unused")
 public class SchedulerService
-  implements ISchedulerService, InitializingBean
+  implements ISchedulerService
 {
   /* Logger */
   private static final Logger logger = LoggerFactory.getLogger(SchedulerService.class);
@@ -57,7 +57,8 @@ public class SchedulerService
   /**
    * The Spring application context.
    */
-  private final ApplicationContext applicationContext;
+  @Autowired
+  private ApplicationContext applicationContext;
 
   /*
    * The delay in milliseconds between successive attempts to execute a job.
@@ -74,31 +75,9 @@ public class SchedulerService
   /**
    * The data source used to provide connections to the application database.
    */
-  private final DataSource dataSource;
-
-  /**
-   * Constructs a new <code>SchedulerService</code>.
-   *
-   * @param applicationContext the Spring application context
-   * @param dataSource         the data source used to provide connections to the application
-   *                           database
-   */
   @Autowired
-  public SchedulerService(ApplicationContext applicationContext, @Qualifier(
-      "applicationDataSource") DataSource dataSource)
-  {
-    this.applicationContext = applicationContext;
-    this.dataSource = dataSource;
-  }
-
-  /**
-   * Initialize the Scheduler Service.
-   */
-  @Override
-  public void afterPropertiesSet()
-  {
-    logger.info(String.format("Initialising the Scheduler Service (%s)", instanceName));
-  }
+  @Qualifier("applicationDataSource")
+  private DataSource dataSource;
 
   /**
    * Create the job.
@@ -107,8 +86,13 @@ public class SchedulerService
    */
   @Override
   public void createJob(Job job)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (job == null)
+    {
+      throw new InvalidArgumentException("job");
+    }
+
     String createJobSQL =
         "INSERT INTO scheduler.jobs (id, name, scheduling_pattern, job_class, is_enabled, status) "
         + "VALUES (?, ?, ?, ?, ?, ?)";
@@ -143,8 +127,13 @@ public class SchedulerService
    */
   @Override
   public void createJobParameter(JobParameter jobParameter)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobParameter == null)
+    {
+      throw new InvalidArgumentException("jobParameter");
+    }
+
     String createJobParameterSQL =
         "INSERT INTO scheduler.job_parameters (id, job_id, name, value) VALUES (?, ?, ?, ?)";
 
@@ -178,8 +167,13 @@ public class SchedulerService
    */
   @Override
   public void deleteJob(UUID jobId)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String deleteJobSQL = "DELETE FROM scheduler.jobs WHERE id=?";
 
     try (Connection connection = dataSource.getConnection();
@@ -206,8 +200,13 @@ public class SchedulerService
    */
   @Override
   public void executeJob(Job job)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (job == null)
+    {
+      throw new InvalidArgumentException("job");
+    }
+
     Class<?> jobClass;
 
     // Load the job class.
@@ -228,7 +227,7 @@ public class SchedulerService
     try
     {
       // Create a new instance of the job
-      Object jobObject = jobClass.getConstructor().newInstance();
+      Object jobObject = jobClass.newInstance();
 
       // Check if the job is a valid job
       if (!(jobObject instanceof IJob))
@@ -284,8 +283,13 @@ public class SchedulerService
    */
   @Override
   public List<Job> getFilteredJobs(String filter)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (filter == null)
+    {
+      throw new InvalidArgumentException("filter");
+    }
+
     String getJobsSQL =
         "SELECT id, name, scheduling_pattern, job_class, is_enabled, status, execution_attempts, "
         + "lock_name, last_executed, next_execution, updated FROM scheduler.jobs";
@@ -324,8 +328,13 @@ public class SchedulerService
    */
   @Override
   public Job getJob(UUID jobId)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String getJobSQL =
         "SELECT id, name, scheduling_pattern, job_class, is_enabled, status, execution_attempts, "
         + "lock_name, last_executed, next_execution, updated FROM scheduler.jobs WHERE id = ?";
@@ -349,8 +358,7 @@ public class SchedulerService
     }
     catch (Throwable e)
     {
-      throw new SchedulerServiceException(String.format("Failed to retrieve the job (%s)", jobId),
-          e);
+      throw new SchedulerServiceException(String.format("Failed to retrieve the job (%s)", jobId), e);
     }
   }
 
@@ -363,8 +371,13 @@ public class SchedulerService
    */
   @Override
   public List<JobParameter> getJobParameters(UUID jobId)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String getJobParametersSQL =
         "SELECT id, job_id, name, value FROM scheduler.job_parameters WHERE job_id = ?";
 
@@ -512,8 +525,13 @@ public class SchedulerService
    */
   @Override
   public int getNumberOfFilteredJobs(String filter)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (filter == null)
+    {
+      throw new InvalidArgumentException("filter");
+    }
+
     String getNumberOfJobsSQL = "SELECT COUNT(id) FROM scheduler.jobs";
 
     String getNumberOfFilteredJobsSQL = "SELECT COUNT(id) FROM scheduler.jobs "
@@ -623,8 +641,13 @@ public class SchedulerService
    */
   @Override
   public void incrementJobExecutionAttempts(UUID jobId)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String incrementJobExecutionAttemptsSQL = "UPDATE scheduler.jobs "
         + "SET execution_attempts=execution_attempts + 1, updated=?, last_executed=? WHERE id=?";
 
@@ -652,6 +675,15 @@ public class SchedulerService
   }
 
   /**
+   * Initialise the Scheduler Service.
+   */
+  @PostConstruct
+  public void init()
+  {
+    logger.info(String.format("Initialising the Scheduler Service (%s)", instanceName));
+  }
+
+  /**
    * Lock a job.
    *
    * @param jobId  the Universally Unique Identifier (UUID) used to uniquely identify the job
@@ -659,8 +691,13 @@ public class SchedulerService
    */
   @Override
   public void lockJob(UUID jobId, JobStatus status)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String lockJobSQL = "UPDATE scheduler.jobs SET status=?, lock_name=?, updated=? WHERE id=?";
 
     try (Connection connection = dataSource.getConnection();
@@ -694,8 +731,13 @@ public class SchedulerService
    */
   @Override
   public void rescheduleJob(UUID jobId, String schedulingPattern)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     try (Connection connection = dataSource.getConnection())
     {
       Predictor predictor = new Predictor(schedulingPattern, System.currentTimeMillis());
@@ -721,7 +763,7 @@ public class SchedulerService
    */
   @Override
   public int resetJobLocks(JobStatus status, JobStatus newStatus)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
     String resetJobLocksSQL = "UPDATE scheduler.jobs SET status=?, lock_name=NULL, updated=? "
         + "WHERE lock_name=? AND status=?";
@@ -829,8 +871,13 @@ public class SchedulerService
    */
   @Override
   public void setJobStatus(UUID jobId, JobStatus status)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     try (Connection connection = dataSource.getConnection())
     {
       setJobStatus(connection, jobId, status);
@@ -850,8 +897,13 @@ public class SchedulerService
    */
   @Override
   public void unlockJob(UUID jobId, JobStatus status)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (jobId == null)
+    {
+      throw new InvalidArgumentException("jobId");
+    }
+
     String unlockJobSQL =
         "UPDATE scheduler.jobs SET status=?, updated=?, lock_name=NULL WHERE id=?";
 
@@ -883,8 +935,13 @@ public class SchedulerService
    */
   @Override
   public void updateJob(Job job)
-    throws SchedulerServiceException
+    throws InvalidArgumentException, SchedulerServiceException
   {
+    if (job == null)
+    {
+      throw new InvalidArgumentException("job");
+    }
+
     String updateJobSQL =
         "UPDATE scheduler.jobs SET name=?, scheduling_pattern=?, job_class=?, is_enabled=?, "
         + "status=? WHERE id=?";
